+++ core.js
@@ -1,4 +1,3 @@
-(function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
 var server = require('./server.js');
 var url = server.url;
 
@@ -157,92 +156,3 @@ pybossa.setEndpoint = function (endpoint) {
 }
 
 module.exports = pybossa;
-
-},{"./server.js":3}],2:[function(require,module,exports){
-// pybossa.js library
-// Copyright (C) 2013 SF Isle of Man
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-
-(function(global) {
-    var pybossa = require('./core.js');
-    global.pybossa = pybossa;
-} (window.pybossa = window.pybossa || {}, jQuery));
-
-},{"./core.js":1}],3:[function(require,module,exports){
-var url = '/';
-
-function _userProgress(projectname) {
-    return $.ajax({
-        url: url + 'api/project/' + projectname + '/userprogress',
-        dataType: 'json'
-    });
-}
-
-function _fetchProject(projectname) {
-    return $.ajax({
-        url: url + 'api/project',
-        data: 'short_name='+projectname,
-        dataType:'json'
-    });
-}
-
-function _fetchNewTask(projectId, offset) {
-    offset = offset || 0;
-    return $.ajax({
-        url: url + 'api/project/' + projectId + '/newtask',
-        data: 'offset=' + offset,
-        dataType: 'json'
-    });
-}
-function _fetchTask(taskId) {
-    return $.ajax({
-        url: url + 'api/task/' + taskId,
-        dataType: 'json'
-    });
-}
-
-function _saveTaskRun(taskrun) {
-    return $.ajax({
-        type: 'POST',
-        url: url + 'api/taskrun',
-        dataType: 'json',
-        contentType: 'application/json',
-        data: taskrun
-    });
-}
-
-function _setEndpoint(endpoint) {
-    // Check that the URL has the trailing slash, otherwise add it
-    if ( endpoint.charAt(endpoint.length-1) != '/' ) {
-        endpoint += '/';
-    }
-    url = endpoint;
-    return url;
-}
-
-var serverQueries = {
-    userProgress: _userProgress,
-    fetchProject: _fetchProject,
-    fetchNewTask: _fetchNewTask,
-    fetchTask:    _fetchTask,
-    saveTaskRun: _saveTaskRun,
-    setEndpoint: _setEndpoint,
-    url:         url
-};
-
-module.exports = serverQueries;
-
-},{}]},{},[2]);
